requirements.txt
bcrypt==4.1.2
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.19.1
fsspec==2025.9.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-generativeai==0.7.2
hf-xet==1.1.10
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.4.2
packaging==26.0
pandas==2.3.3
PyPDF2==3.0.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
sentence-transformers==2.3.1
scikit-learn==1.8.0
streamlit==1.53.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
python-dotenv==1.2.1
